requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.2
opencv-python-headless==4.8.0.74
boto3==1.28.8
Pillow==10.0.1
requests==2.31.0
numpy==1.25.2
mediapipe==0.10.3
python-socketio==5.8.0
eventlet==0.33.3
gunicorn